Integrate destruction effects into group manager
