--- Runtime/Effects.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 260033821b07c9847b3a5a406c2b4e4d
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
