--- Runtime/Effects/DestructionEffects.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 5a417e11780088841b25deb0aeb7fcd2
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
